+++ src/engine/Code/include/autosql/tests/test14_numerical_output.hpp
@@ -56,7 +56,7 @@ void test14_numerical_output()
 
     for ( size_t i = 0; i < join_keys_population.size(); ++i )
         {
-            join_keys_population[i] = i;
+            join_keys_population[i] = static_cast<std::int32_t>( i );
         }
 
     const auto join_keys_population_col =
