+++ src/engine/Code/include/autosql/tests/test21_snowflake_model.hpp
@@ -108,7 +108,7 @@ void test21_snowflake_model()
 
     for ( size_t i = 0; i < join_keys_population.size(); ++i )
         {
-            join_keys_population[i] = i;
+            join_keys_population[i] = static_cast<std::int32_t>( i );
         }
 
     const auto join_keys_population_col =
