+++ src/engine/Code/src/relboost/DecisionTreeEnsemble.cpp
@@ -57,7 +57,7 @@ DecisionTreeEnsemble::DecisionTreeEnsemble(
     for ( size_t i = 0; i < peripheral_arr.size(); ++i )
         {
             peripheral.push_back(
-                containers::Schema( *peripheral_arr.getObject( i ) ) );
+                containers::Schema( *peripheral_arr.getObject( static_cast<unsigned int>( i ) ) ) );
         }
 
     impl().peripheral_schema_.reset(
@@ -73,7 +73,7 @@ DecisionTreeEnsemble::DecisionTreeEnsemble(
                 _encoding,
                 impl().hyperparameters_,
                 loss_function_,
-                *trees_objects.getObject( i ) ) );
+                *trees_objects.getObject( static_cast<unsigned int>( i ) ) ) );
         }
 
     // ------------------------------------------------------------------------
